autoflow/requirements.txt
python-dotenv==1.0.0
supabase
yfinance==0.2.36
pandas==2.2.0
python-telegram-bot==20.7
aiohttp==3.9.3
asyncio==3.4.3
httpx>=0.27.0 